:twisted_rightwards_arrows: Fix AI category validation and build warnings (#1376)
* :bug: (ai): constrain category recognition to synced categories

* :bug: (tap): migrate TensorFlow Lite runtime to LiteRT

* :rotating_light: (build): clean up Gradle and compiler warnings

* :bug: (ai): reject ambiguous child category mappings

* :bug: (server): avoid duplicate local server initialization

* :rotating_light: (build): address explicit lint findings

+++ tap/src/main/java/net/ankio/tap/samsung/SamsungBackTapDetectionService.kt
@@ -66,7 +66,7 @@ class SamsungBackTapDetectionService(
                 lastT,
                 sSamplingIntervalInNano,
             )
-            val v = Math.toDegrees(acos((lastZ / f).toDouble()).toDouble()).toInt()
+            val v = Math.toDegrees(acos((lastZ / f).toDouble())).toInt()
             mIsFlat[mFlatIndex] = if (v < 10 || v > 170) 1 else 0
             val v1 = mFlatIndex + 1
             mFlatIndex = v1
